Add files via upload
"Amazon Clone Landing Page HTML/CSS Template"

This GitHub repository contains a simplified HTML/CSS template that closely resembles the landing page design of Amazon.com. The template includes various sections of the Amazon homepage, such as the navigation bar, search bar, hero section, shop section, and footer. It utilizes SCSS (Sass) for enhanced styling and features some interactive elements like hover effects.

Please note that this is not a functional e-commerce website but rather a static representation of Amazon's landing page for learning and demonstration purposes. The template is intended to showcase frontend web development skills and can serve as a starting point for building more complex projects.

As this is only a static template, it does not include any backend functionality or dynamic features like user authentication, shopping cart functionality, or actual product listings. Developers can fork and customize this template according to their needs, add JavaScript functionality, and connect it to a backend to create a fully functional e-commerce site.

Disclaimer: This project is created for educational purposes and is not an official Amazon product or endorsed by Amazon.com. The use of Amazon's logo and design elements is purely for imitation and demonstration, and users are advised to respect copyright and trademark laws while using this template